update game page metadata
