+++ web/src/app/layout.tsx
@@ -9,8 +9,19 @@ import './globals.css';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-    title: 'PlayBingo',
+    title: { template: '%s | PlayBingo', default: 'PlayBingo' },
     description: 'The next generation of gaming bingo',
+    openGraph: {
+        title: {
+            template: '%s | PlayBingo',
+            default: 'PlayBingo',
+        },
+        images: {
+            url: 'https://playbingo.gg/logo.png',
+            width: 5900,
+            height: 1867,
+        },
+    },
 };
 
 export default function RootLayout({
